Rename method to avoid conflict with reserved Windows function name
Changed the name of the method `ShowWindow` to avoid the error "conflicting declaration of C function 'void ShowWindow(int)'". This error occurred because `ShowWindow` is a reserved function name in Windows.Rename method to avoid conflict with reserved Windows function name

Changed the name of the method `ShowWindow` to avoid the error "conflicting declaration of C function 'void ShowWindow(int)'". This error occurred because `ShowWindow` is a reserved function name in Windows.